correção cadastro agenda com o usuário
