+++ src/main/java/br/com/deveficiente/calendario/controller/UsuarioController.java
@@ -1,7 +1,9 @@
 package br.com.deveficiente.calendario.controller;
 
 import br.com.deveficiente.calendario.controller.form.UsuarioForm;
+import br.com.deveficiente.calendario.model.Agenda;
 import br.com.deveficiente.calendario.model.Usuario;
+import br.com.deveficiente.calendario.repository.AgendaRepository;
 import br.com.deveficiente.calendario.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -19,13 +21,16 @@ public class UsuarioController {
 
     @Autowired
     private UsuarioRepository usuarioRepository;
+    @Autowired
+    private AgendaRepository agendaRepository;
 
     @Transactional
     @PostMapping
     public ResponseEntity cadastrar(@RequestBody @Valid UsuarioForm usuarioForm) {
         Usuario usuario = usuarioForm.converter();
-
-        usuarioRepository.save(usuario);
+        usuario = usuarioRepository.save(usuario);
+        Agenda agenda = new Agenda(usuario.getNome(), "Agenda "+ usuario.getNome(), usuario);
+        agendaRepository.save(agenda);
         return ResponseEntity.ok("Usuário  cadastrado");
 
     }
