+++ src/main/java/br/com/deveficiente/calendario/controller/form/UsuarioForm.java
@@ -1,19 +1,16 @@
 package br.com.deveficiente.calendario.controller.form;
 
 import br.com.deveficiente.calendario.model.Usuario;
-import lombok.Data;
 import org.apache.commons.codec.digest.DigestUtils;
-import org.apache.commons.codec.digest.MessageDigestAlgorithms;
 import org.hibernate.validator.constraints.Length;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
-import java.security.MessageDigest;
-import java.util.Date;
 
 public class UsuarioForm {
-
+    @NotNull @NotEmpty
+    private String nome;
     @NotNull @NotEmpty @Email(message = "O login deve ser um email válido")
     private String login;
     @NotNull @NotEmpty @Length(min = 6, message = "A senha deve ter no mínimo 6 caracteres")
@@ -21,7 +18,15 @@ public class UsuarioForm {
 
     public Usuario converter(){
         String senhaCriptografada = DigestUtils.md5Hex(senha.getBytes()).toUpperCase();
-        return new Usuario(login, senhaCriptografada);
+        return new Usuario(nome,login, senhaCriptografada);
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
     }
 
     public String getLogin() {
