+++ src/main/java/br/com/deveficiente/calendario/model/Usuario.java
@@ -1,8 +1,9 @@
 package br.com.deveficiente.calendario.model;
 
-import lombok.Data;
-
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import javax.validation.constraints.Pattern;
 import java.time.LocalDateTime;
 
@@ -11,13 +12,15 @@ public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+    private String nome;
     private String login;
     @Pattern(regexp = "^[a-fA-F0-9]{32}$", message = "A senha deve ser criptografada em MD5")
     private String senha;
     private LocalDateTime dataCadastro;
 
 
-    public Usuario(String login, String senha) {
+    public Usuario(String nome, String login, String senha) {
+        this.nome = nome;
         this.login = login;
         this.senha = senha;
         this.dataCadastro = LocalDateTime.now();
@@ -29,6 +32,14 @@ public class Usuario {
 
     public void setId(Long id) {
         this.id = id;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
     }
 
     public String getLogin() {
