--- src/main/java/br/com/deveficiente/calendario/repository/PessoaRepository.java
@@ -1,7 +1,0 @@
-package br.com.deveficiente.calendario.repository;
-
-import br.com.deveficiente.calendario.model.Pessoa;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PessoaRepository extends JpaRepository<Pessoa, Long> {
-}
