+++ src/main/resources/application.properties
@@ -10,3 +10,4 @@ spring.jpa.hibernate.ddl-auto=create
 # h2
 spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
+spring.h2.m
